requirements.txt
# Python dependencies for the Vault proof‑of‑concept
# hvac is the official Python client library for HashiCorp Vault. It wraps
# the HTTP API and makes it easy to interact with Vault from Python.
hvac>=1.3.0

# python‑dotenv is used to load environment variables from a .env file. This
# project optionally supports reading the VAULT_ADDR and VAULT_TOKEN
# environment variables from a .env file placed in the project root.
python-dotenv>=1.0.0
